Add EventController, routes, some adjusments in functionality and testing

+++ src/services/EventService.ts
@@ -1,6 +1,7 @@
 import { EventType, IEvent } from "../models/Event";
 import EventRepository from "../repositories/EventRepository";
 import validateDate from "../utils/date-validator";
+import { Types } from "mongoose";
 
 export type EventData = {
   title: string;
@@ -15,12 +16,10 @@ export default class EventService {
   }
 
   public static async getEventById(id: string): Promise<IEvent | null> {
-    try {
+    if (Types.ObjectId.isValid(id)) {
       return EventRepository.getEventById(id);
-    } catch (err) {
-      console.error(err);
-      return null;
     }
+    return null;
   }
 
   public static async getEventBytitle(title: string): Promise<IEvent | null> {
@@ -33,12 +32,17 @@ export default class EventService {
       return Promise.reject("The date is invalid!");
     }
     const { title, description, createdBy } = event;
+    const existingEvent = await EventRepository.getEventBytitle(title);
+    if (existingEvent && existingEvent.createdBy === createdBy)
+      return Promise.reject("Duplicate titles are not allowed!");
 
     const eventsPerDay = (
       await EventRepository.getUserEventsByDate(createdBy, date)
     ).length;
-    if (eventsPerDay >= 5)
+
+    if (eventsPerDay >= 5) {
       return Promise.reject("User can't create more than 5 events per day!");
+    }
 
     const newEvent: EventType = {
       title,
@@ -52,12 +56,20 @@ export default class EventService {
   }
 
   public static async updateEvent(id: string, data: Record<string, any>) {
-    if (!data.date) {
-      return EventRepository.updateEvent(id, data);
+    if (!Types.ObjectId.isValid(id)) {
+      return Promise.reject("Id doesn't exist!");
     }
+    const event = await EventRepository.getEventById(id);
     const date = validateDate(data.date);
     if (!date) {
-      return null;
+      return Promise.reject("Date must be in the future!");
+    }
+    const eventsPerDay = (
+      await EventRepository.getUserEventsByDate(event.createdBy, date)
+    ).length;
+
+    if (eventsPerDay >= 5) {
+      return Promise.reject("User can't create more than 5 events per day!");
     }
 
     return EventRepository.updateEvent(id, data);
@@ -65,18 +77,19 @@ export default class EventService {
 
   public static async registerUserToEvent(event_id: string, username: string) {
     let event;
-    try {
+    if (Types.ObjectId.isValid(event_id)) {
       event = await EventRepository.getEventById(event_id);
-    } catch (err) {
-      return null;
     }
-    if (!event) return null;
+    if (!username) return Promise.reject("username can't be null!");
+
+    if (!event) return Promise.reject("Event doesn't exist!");
 
     if (event.users.get(username))
       return Promise.reject("User already registered!");
 
-    if (!validateDate(event.date.toString()))
-      return Promise.reject("The event already passed!");
+    const now = new Date();
+
+    if (event.date < now) return Promise.reject("The event already passed!");
 
     const users = event.users;
     users.set(username, true);
@@ -84,11 +97,7 @@ export default class EventService {
   }
 
   public static async deleteEvent(id: string): Promise<IEvent | null> {
-    try {
-      return EventRepository.deleteEvent(id);
-    } catch (err) {
-      console.error(err);
-      return null;
-    }
+    if (!Types.ObjectId.isValid(id)) return null;
+    return EventRepository.deleteEvent(id);
   }
 }
